pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-smart_router"
version = "0.1.0"
description = "Smart Router environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    # install from github
    # "openenv-core[core] @ git+https://github.com/meta-pytorch/OpenEnv.git",
    "openenv-core[core]>=0.2.2",
    # LLM backends for judge/curriculum
    "openai>=1.0.0",
    "anthropic>=0.39.0",
    "huggingface_hub>=0.20.0",
    # Visualization
    "matplotlib>=3.7.0",
    # Environment-specific dependencies
    # Add all dependencies needed for your environment here
    # Examples:
    # "numpy>=1.19.0",
    # "torch>=2.0.0",
    # "gymnasium>=0.29.0",
    # "openspiel>=1.0.0",
    # "smolagents>=1.22.0,<2",
]

[project.optional-dependencies]
train = [
    "trl[vllm]==0.29.0",
    "vllm==0.11.2",
    "torch>=2.8.0,<2.10.0",
    "peft",
    "transformers",
    "datasets",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]
eval = [
    "vllm==0.11.2",
    "transformers",
    "tabulate>=0.9.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m smart_router.server.app
server = "smart_router.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["smart_router", "smart_router.server"]
package-dir = { "smart_router" = ".", "smart_router.server" = "server" }